+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The following is a history of the changes made to this project.
 * The dump978 has been rewritten in order to use FlightAwares version of dump978.
 * Removed the option to install dump1090-mutability.
 * Removed the option to install dump1090 HPTOA.
+* Removed the option to install AboveTustin.
 * Removed incomplete RTL-SDR OGN setup scripting.
 * Version check for dump1090-fa has been fixed.
 * Fix bug where PiAware was installed every run even if already installed.